+++ src/app/core/services/auth/auth.constants.ts
@@ -1,0 +1,9 @@
+/**
+ * Copyright (c) 2025, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+/** URL of the Apache CGI endpoint that returns OIDC claims (PKCE-protected). */
+export const USERINFO_URL = '/auth/userinfo';
